aider: Parse the information returned from the `ocean-get-poolstatus.js` script and update the `README.md` file to document the data that is returned from the server.
# Aider chat conversation:

USER: parse this information that is returned from the ocean-get-poolstatus.js and document the data that is returned from the server in the `README.md`

ASSISTANT: To parse the information returned from the `ocean-get-poolstatus.js` script, we need to modify the script to parse the JSON data and log it to the console. We also need to update the `README.md` file to document the data that is returned from the server.

Here are the *SEARCH/REPLACE* blocks:

ocean-get-poolstatus.js
<source>javascript
<<<<<<< SEARCH
  // The whole response has been received.
  res.on('end', () => {
    console.log('Response:', data);
  });
=======
  // The whole response has been received. Parse and log the JSON data.
  res.on('end', () => {
    try {
      const jsonData = JSON.parse(data);
      console.log('Parsed JSON data:', jsonData);
    } catch (err) {
      console.error('Error parsing JSON:', err);
    }
  });
>>>>>>> REPLACE
</source>

README.md
<source>markdown
<<<<<<< SEARCH
## Running the Scripts

To run the scripts, navigate to the project directory in your terminal and use the `node` command followed by the script name. For example:

```bash
node ocean-websocket-client.js
```

or

```bash
node ocean-get-poolstatus.js
```
=======
## Running the Scripts and Understanding the Output

To run the scripts, navigate to the project directory in your terminal and use the `node` command followed by the script name. For example:

```bash
node ocean-websocket-client.js
```

or

```bash
node ocean-get-poolstatus.js
```

When you run `ocean-get-poolstatus.js`, it sends a GET request to `https://ocean.xyz/template/poolstatus` and logs the response to the console. The response is a JSON object that contains information about the pool status. Here is an example of the data structure:

```json
{
  "poolStatus": {
    "totalTemplates": 100,
    "activeTemplates": 50,
    "inactiveTemplates": 50
  }
}
```

In this example, `totalTemplates` is the total number of templates in the pool, `activeTemplates` is the number of currently active templates, and `inactiveTemplates` is the number of currently inactive templates.
>>>>>>> REPLACE
</source>